requirements.txt
pymol
gmxapi
matplotlib
gromacs-wrapper
gromacs 2020 beta2
